backend/requirements.txt
Django==5.1.5
Requests==2.32.3
django-redis==5.4.0
python-decouple==3.8
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
psycopg2-binary